requirements.txt
streamlit==1.12.0
pandas==1.3.5
numpy==1.26.4
plotly==5.5.0
streamlit-option-menu==0.3.2
numerize==0.10.0
mysql-connector-python==9.0.0

